fixes build issues
