+++ integration/tests/common/receiptstate/receipt_state.go
@@ -8,13 +8,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/onflow/flow-go/integration/tests/common/blockstate"
 	"github.com/onflow/flow-go/model/flow"
-	"github.com/onflow/flow-go/model/messages"
-	"github.com/onflow/flow-go/utils/unittest"
 )
 
-const receiptStateTimeout = 60 * time.Second
+const Timeout = 60 * time.Second
 
 type ReceiptState struct {
 	sync.RWMutex
@@ -48,9 +45,9 @@ func (rs *ReceiptState) WaitForReceiptFromAny(t *testing.T, blockID flow.Identif
 		defer rs.RUnlock()
 
 		return len(rs.receipts[blockID]) > 0
-	}, receiptStateTimeout, 100*time.Millisecond,
+	}, Timeout, 100*time.Millisecond,
 		fmt.Sprintf("did not receive execution receipt for block ID %x from any node within %v seconds", blockID,
-			receiptStateTimeout))
+			Timeout))
 	for _, r := range rs.receipts[blockID] {
 		return r
 	}
@@ -67,71 +64,8 @@ func (rs *ReceiptState) WaitForReceiptFrom(t *testing.T, blockID, executorID flo
 		var ok bool
 		r, ok = rs.receipts[blockID][executorID]
 		return ok
-	}, receiptStateTimeout, 100*time.Millisecond,
+	}, Timeout, 100*time.Millisecond,
 		fmt.Sprintf("did not receive execution receipt for block ID %x from %x within %v seconds", blockID, executorID,
-			receiptStateTimeout))
+			Timeout))
 	return r
-}
-
-// WaitUntilFinalizedStateCommitmentChanged waits until a different state commitment for a finalized block is received
-// compared to the latest one from any execution node and returns the corresponding block and execution receipt
-func WaitUntilFinalizedStateCommitmentChanged(t *testing.T, bs *blockstate.BlockState, rs *ReceiptState,
-	qualifiers ...func(receipt flow.ExecutionReceipt) bool) (*messages.BlockProposal,
-	*flow.ExecutionReceipt) {
-
-	// get the state commitment for the highest finalized block
-	initialFinalizedSC := unittest.GenesisStateCommitment
-	var err error
-	b1, ok := bs.HighestFinalized()
-	if ok {
-		r1 := rs.WaitForReceiptFromAny(t, b1.Header.ID())
-		initialFinalizedSC, err = r1.ExecutionResult.FinalStateCommitment()
-		require.NoError(t, err)
-	}
-
-	initFinalizedheight := b1.Header.Height
-	currentHeight := initFinalizedheight + 1
-
-	currentID := b1.Header.ID()
-	var b2 *messages.BlockProposal
-	var r2 *flow.ExecutionReceipt
-	require.Eventually(t, func() bool {
-		var ok bool
-		b2, ok = bs.FinalizedHeight(currentHeight)
-		if !ok {
-			return false
-		}
-		currentID = b2.Header.ID()
-		r2 = rs.WaitForReceiptFromAny(t, b2.Header.ID())
-		r2finalState, err := r2.ExecutionResult.FinalStateCommitment()
-		require.NoError(t, err)
-		if initialFinalizedSC == r2finalState {
-			// received a new execution result for the next finalized block, but it has the same final state commitment
-			// check the next finalized block
-			currentHeight++
-			return false
-		}
-
-		for _, qualifier := range qualifiers {
-			if !qualifier(*r2) {
-				return false
-			}
-		}
-
-		return true
-	}, receiptStateTimeout, 100*time.Millisecond,
-		fmt.Sprintf("did not receive an execution receipt with a different state commitment from %x within %v seconds,"+
-			" initial finalized height: %v "+
-			" last block checked height %v, last block checked ID %x", initialFinalizedSC, receiptStateTimeout,
-			initFinalizedheight,
-			currentHeight, currentID))
-
-	return b2, r2
-}
-
-// WithMinimumChunks creates a qualifier that returns true if receipt has the specified minimum number of chunks.
-func WithMinimumChunks(chunkNum int) func(flow.ExecutionReceipt) bool {
-	return func(receipt flow.ExecutionReceipt) bool {
-		return len(receipt.ExecutionResult.Chunks) >= chunkNum
-	}
 }
