+++ integration/tests/common/testnet_state_tracker.go
@@ -2,6 +2,7 @@ package common
 
 import (
 	"context"
+	"fmt"
 	"strings"
 	"testing"
 	"time"
@@ -13,6 +14,7 @@ import (
 	"github.com/onflow/flow-go/integration/tests/common/receiptstate"
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/model/messages"
+	"github.com/onflow/flow-go/utils/unittest"
 )
 
 type TestnetStateTracker struct {
@@ -105,4 +107,67 @@ func (tst *TestnetStateTracker) Track(t *testing.T, ctx context.Context, ghost *
 			}
 		}
 	}()
+}
+
+// WaitUntilFinalizedStateCommitmentChanged waits until a different state commitment for a finalized block is received
+// compared to the latest one from any execution node and returns the corresponding block and execution receipt
+func WaitUntilFinalizedStateCommitmentChanged(t *testing.T, bs *blockstate.BlockState, rs *receiptstate.ReceiptState,
+	qualifiers ...func(receipt flow.ExecutionReceipt) bool) (*messages.BlockProposal,
+	*flow.ExecutionReceipt) {
+
+	// get the state commitment for the highest finalized block
+	initialFinalizedSC := unittest.GenesisStateCommitment
+	var err error
+	b1, ok := bs.HighestFinalized()
+	if ok {
+		r1 := rs.WaitForReceiptFromAny(t, b1.Header.ID())
+		initialFinalizedSC, err = r1.ExecutionResult.FinalStateCommitment()
+		require.NoError(t, err)
+	}
+
+	initFinalizedheight := b1.Header.Height
+	currentHeight := initFinalizedheight + 1
+
+	currentID := b1.Header.ID()
+	var b2 *messages.BlockProposal
+	var r2 *flow.ExecutionReceipt
+	require.Eventually(t, func() bool {
+		var ok bool
+		b2, ok = bs.FinalizedHeight(currentHeight)
+		if !ok {
+			return false
+		}
+		currentID = b2.Header.ID()
+		r2 = rs.WaitForReceiptFromAny(t, b2.Header.ID())
+		r2finalState, err := r2.ExecutionResult.FinalStateCommitment()
+		require.NoError(t, err)
+		if initialFinalizedSC == r2finalState {
+			// received a new execution result for the next finalized block, but it has the same final state commitment
+			// check the next finalized block
+			currentHeight++
+			return false
+		}
+
+		for _, qualifier := range qualifiers {
+			if !qualifier(*r2) {
+				return false
+			}
+		}
+
+		return true
+	}, receiptstate.Timeout, 100*time.Millisecond,
+		fmt.Sprintf("did not receive an execution receipt with a different state commitment from %x within %v seconds,"+
+			" initial finalized height: %v "+
+			" last block checked height %v, last block checked ID %x", initialFinalizedSC, receiptstate.Timeout,
+			initFinalizedheight,
+			currentHeight, currentID))
+
+	return b2, r2
+}
+
+// WithMinimumChunks creates a qualifier that returns true if receipt has the specified minimum number of chunks.
+func WithMinimumChunks(chunkNum int) func(flow.ExecutionReceipt) bool {
+	return func(receipt flow.ExecutionReceipt) bool {
+		return len(receipt.ExecutionResult.Chunks) >= chunkNum
+	}
 }
